+++ spec/rails_spec.rb
@@ -13,28 +13,28 @@ describe "Cucumba[:server]" do
     Dir[File.join('app/models/*.rb')].each { |model| load model }
   end
 
-  it "should return instance of Cucumba::Rails" do
-    Cucumba[:test].should be_instance_of(Cucumba::Rails)
+  it "should return instance of Cucumba::Railz" do
+    Cucumba[:test].should be_instance_of(Cucumba::Railz)
   end
 
-  context "#model" do
+  context "#class" do
 
-    it "should return instace of Cucumba::Rails::Model" do
-      [ Cucumba[:test].model("User"), Cucumba[:test].m("User"), Cucumba[:test].model(:User), Cucumba[:test].m(:User) ].each do |model|
-	model.should be_instance_of(Cucumba::Rails::Model)
+    it "should return instace of Cucumba::Railz::Klass" do
+      [ Cucumba[:test].class("User"), Cucumba[:test].c("User"), Cucumba[:test].class(:User), Cucumba[:test].c(:User) ].each do |object|
+	object.should be_instance_of(Cucumba::Railz::Klass)
       end
     end
 
     it "should invoke existing methods" do
       users_count_was = User.all.count
-      Cucumba[:test].m("User").create!(:name => 'test user', :password => 'secret')
+      Cucumba[:test].c("User").create!(:name => 'test user', :password => 'secret')
       (User.all.count - users_count_was).should == 1
       User.last.name.should == 'test user'
       User.last.password.should == 'secret'
     end
 
     it "should invoke method without arguments" do
-      Cucumba[:test].m(:User).columns
+      Cucumba[:test].c(:User).columns
     end
 
     it "should eval code in server side" do
@@ -48,19 +48,19 @@ describe "Cucumba[:server]" do
     end
 
     it "should invoke this method" do
-      Cucumba[:test].m(:User).method_with_args(:name, :first => 'bla', :last => 'alb')
+      Cucumba[:test].c(:User).method_with_args(:name, :first => 'bla', :last => 'alb')
     end
 
-    it "should raise exception if model does not exists" do
-      lambda { Cucumba[:test].m("Unknown") }.should raise_exception(Cucumba::Rails::Model::NotFoundError)
+    it "should raise exception if class does not exists" do
+      lambda { Cucumba[:test].c("Unknown") }.should raise_exception(Cucumba::Railz::Klass::NotFoundError)
     end
 
     it "should raise exception if model's method does not exists" do
-      lambda { Cucumba[:test].m(:User).blabla }.should raise_exception(NoMethodError)
+      lambda { Cucumba[:test].c(:User).blabla }.should raise_exception(NoMethodError)
     end
 
     it "should raise exception if exception appears on server" do
-      lambda { Cucumba[:test].m(:User).create!(:password => 'secret') }.should raise_exception(RuntimeError,"ActiveRecord::RecordInvalid Validation failed: Name can't be blank")
+      lambda { Cucumba[:test].c(:User).create!(:password => 'secret') }.should raise_exception(RuntimeError,"ActiveRecord::RecordInvalid Validation failed: Name can't be blank")
     end
 
   end